Split session adapter code into separate files from the analyzers
