+++ src/packet_analysis/protocol/icmp/Plugin.cc
@@ -3,6 +3,7 @@
 #include "zeek/plugin/Plugin.h"
 #include "zeek/packet_analysis/Component.h"
 #include "zeek/packet_analysis/protocol/icmp/ICMP.h"
+#include "zeek/packet_analysis/protocol/icmp/ICMPSessionAdapter.h"
 #include "zeek/analyzer/Component.h"
 
 namespace zeek::plugin::Zeek_ICMP {
