+++ src/packet_analysis/protocol/udp/Plugin.cc
@@ -3,6 +3,7 @@
 #include "zeek/plugin/Plugin.h"
 #include "zeek/packet_analysis/Component.h"
 #include "zeek/packet_analysis/protocol/udp/UDP.h"
+#include "zeek/packet_analysis/protocol/udp/UDPSessionAdapter.h"
 #include "zeek/analyzer/Component.h"
 
 namespace zeek::plugin::Zeek_UDP {
